[#1584] Extract footer component and add it to new-layout
